requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
Werkzeug==3.0.4
google-auth-oauthlib
google-auth
google-api-python-client
